refactor: App to Page directory
