+++ src/main.tsx
@@ -2,9 +2,9 @@ import ReactDOM from 'react-dom/client';
 import { RouterProvider, createBrowserRouter } from 'react-router-dom';
 import { DetailsCard, Error } from './components';
 import './styles/index.css';
-import App from './App';
 import { initFetchData } from './loaders';
 import React from 'react';
+import { App } from './pages';
 
 const router = createBrowserRouter([
   {
