+++ src/pages/index.tsx
@@ -1,0 +1,1 @@
+export { default as App } from './Home/App';
